Refactored Home GUI Config
Dconf is no longer a module. It's now called by the Gnome default.nix. I
also refactored each DE to have a default.nix and install user's gui
apps (rather than doing so from the flake. This provided flexability
when it comes to not installing an applicaiton in some DE while
simplifying the flake.