Utalized sops-nix Home Manager Module
I added the Home Manager version of sops-nix as an import in the
flake.nix in the last commit. Now I've added a sops-nix Home Manager
module to the reedclanton user. I also updated the element-desktop nix
file to do what I want to do with Home Manager, but do it with NixOS
insted (although it's commented out because it forcues the usage of
`--impure` when building).