Removed brightnessctl from desktop Host
brightnessctl doesn't work on my desktop monitors, so I removed it. Now
that my laptop is the only machine with `brightnessctl` installed, I
decided to move the sudo exception from the global sudo module, to the
framework13 one. I also did this because I think it could represent a
secrity threat.