Refactored User Non-HM Flatpaks
Moved non home manager flatpaks so the directory sturcture is the same
as for home manager ones. Removed user install of flatseal but kept
system one. Only one is needed and I figure the system one is more
likely to have more permissions. Also turned on removal of unmanaged
flatpaks. This leads to The Battle for Wesnoth being uninstalled and
then reinstalled. Need to resolve this bug in nix-flatpak. Also removed
direct calls to flatpak from flake.nix because that should be abstracted
out by the default file (and is now).