Organize applications (#11)
* Moved User HM Packages

* Corrected Paths

Corrected paths after moving stuff. Also found typos in paths.

* Moved HM tty Packages

* Moved HM tty programs

* Corrected outdate path

* Moved HM GUI Programs

* Moved HM Flatpaks

* Corrected outdated path

* Refactored User Non-HM Flatpaks

Moved non home manager flatpaks so the directory sturcture is the same
as for home manager ones. Removed user install of flatseal but kept
system one. Only one is needed and I figure the system one is more
likely to have more permissions. Also turned on removal of unmanaged
flatpaks. This leads to The Battle for Wesnoth being uninstalled and
then reinstalled. Need to resolve this bug in nix-flatpak. Also removed
direct calls to flatpak from flake.nix because that should be abstracted
out by the default file (and is now).

* Fixed VM Host issues with recent changes

* Reorganized host specific HM files

The layout now matches the rest of the home manager files.

* Refactored FrameWork13 Host and Common Modules

Refactored the to be in line with home manager config layout.

* Moved Flatpak to Common Modules

Moved flatpak setup and the most basic flatpak apps to common modules
because the setup was almost the same in multiple places.

* Minor Clean up