Move networking into hardware directory (#18)
* desktop host networking into hardware directory

Moving hardware stuff into the hardware directory. This should clean up
the repo and make stuff easier to find.

* User networking into hardware directory

* User desktop host moved networking into hardware

* Improved Resilency

Ensured that when calls to files that don't exist fail, it's ignored.
This is done so parts of the configuration may be removed without
breaking everything.

* Refactored how game controller hardware is called

* Added what's sourced

Done so it's less likely that I'll need to add stuff to this file in the
future.

* Networking move wrap up for desktop

* Tracking current state

* Added conditional import to desktop host

* Framework13 networking move except HM

* VM networking move

* Moved networking in common modules

* Added a bunch of unused imports

If any of the imported files are added in the future, then I won't need
to update the imports.