Added Program List, Refactor Lists
Added lists for NixOS and Home Manager installed programs. Also
refactored lists so you can generate any one of them, or a list with all
NixOS applications installed as well as a list with all Home Manager
ones installed. This still don't include everything, but it does include
a lot.