VMWare Site Down Patch
VMWare site is down for maintenance
(https://store-us.vmware.com/workstation_buy_dual), these changes are to
work around this. They should be reverted when the site is back up.

Also, it's nessiarily to add a file that will be used as the replacement
package at `modules/nixos/virtualisation/vmware.bundle`. This may be
downloaded from
https://www.techspot.com/downloads/downloadnow/189/?evp=f14a48a23bc560f5fbe81b8d83387b41&file=10774