Sys Update, Hardware Refactor, Disabled Virtualiz.
VirtualBox is broken on NixOS. The fix is in work, when I have time. In
the meantime I've disabled all virtualization. This allows me to update
everything.

Bugs were introduced into the nixos-hardware repo, so I switched from
using master to a specific commit.

I updated how hardware configuration is called for Framework13 and
desktop hosts. This was done so a warning is produced when a file that
should exist doesn't. If the file must exist for the system to be build,
then an exception will be raised. This was done so if someone doesn't
want to bother configuring something optional like bluetooth, they don't
have to in order to build the system.

This work has been done on a separate branch from master because I'd like
to be able to revert to a configuration that has VirtualBox on it.