Hotfix/virtualization disabled (#33)
* Sys Update, Hardware Refactor, Disabled Virtualiz.

VirtualBox is broken on NixOS. The fix is in work, when I have time. In
the meantime I've disabled all virtualization. This allows me to update
everything.

Bugs were introduced into the nixos-hardware repo, so I switched from
using master to a specific commit.

I updated how hardware configuration is called for Framework13 and
desktop hosts. This was done so a warning is produced when a file that
should exist doesn't. If the file must exist for the system to be build,
then an exception will be raised. This was done so if someone doesn't
want to bother configuring something optional like bluetooth, they don't
have to in order to build the system.

This work has been done on a separate branch from master because I'd like
to be able to revert to a configuration that has VirtualBox on it.

* Added Nix Bookmarks

* KeePass, Bookmarks, & NeoVIM Spell Dictionary

Updated address of AA FCU and Discover accounts in KeePass. Added FCU
for Federal Credit Unit to NeoVIM spell dictionary, and added bookmarks.

* Updated boot, nix, and NeoVIM Spell Dictionary

Updated boot option that sets number of past generations to keep now
that I have more confidence in this configuration that's it's
up-to-date. Turned back on garbage collection for the same reason.

Also added words to NeoVIM spell dictionary.

* Added Workaround for Broken VMware Package

The URL NixOS uses to download the VMware package no longer works. I've
implemented a workaround while I attempt to create a Nix overlay.

* Game Controller Import Fix

When I changed how game controller configuration was imported, I made
the mistake of assuming that it was imported the same as everything
else. While I maybe should fix it so it does, in the mean time I'm just
going to fix how it's imported.

* Cleaned Up USB Configuration

* Increased Number of System Versions Saved

30 just felt better than 25.